seperates apis and views routes, static files, adds apis' library
Seperates out the routes for APIs and views so they can be easily broken
down into separate frontend and backend applications if needed in
future. Creates static HTML file in seperate directory so it can be
exported to NGINX root to serve instead of service entertaining views
requests. Adds an API library that can have functions related to Solr or
other parts of service.